CPM-1808: Add tests
