+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/attribute/form/hooks/__mocks__/useMainIdentifierCode.ts
@@ -1,0 +1,5 @@
+const useMainIdentifierCode = () => {
+  return 'sku';
+};
+
+export {useMainIdentifierCode};
