+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/attribute/form/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useGetIdentifierAttributesCount';
+export * from './useMainIdentifierCode';
