+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/attribute/form/hooks/useGetIdentifierAttributesCount.ts
@@ -1,3 +1,4 @@
+/* istanbul ignore file */
 import React, {useState} from 'react';
 import {useRouter} from '@akeneo-pim-community/shared';
