+++ src/Akeneo/Platform/Bundle/UIBundle/tests/front/unit/attribute/form/delete/DeleteAction.unit.tsx
@@ -4,6 +4,7 @@ import {fireEvent, screen, getByText, act} from '@testing-library/react';
 import {DeleteAction} from 'pimui/js/attribute/form/delete/DeleteAction';
 import {renderWithProviders} from '@akeneo-pim-community/legacy-bridge/tests/front/unit/utils';
 import {dependencies} from '@akeneo-pim-community/legacy-bridge';
+jest.mock('../../../../../../Resources/public/js/attribute/form/hooks/useMainIdentifierCode');
 
 declare global {
   namespace NodeJS {
