+++ src/Akeneo/Platform/Bundle/UIBundle/tests/front/unit/attribute/form/delete/DeleteModal.unit.tsx
@@ -4,6 +4,7 @@ import {fireEvent, screen, act} from '@testing-library/react';
 import {dependencies} from '@akeneo-pim-community/legacy-bridge';
 import {DeleteModal} from 'pimui/js/attribute/form/delete/DeleteModal';
 import {renderWithProviders} from '@akeneo-pim-community/legacy-bridge/tests/front/unit/utils';
+jest.mock('../../../../../../Resources/public/js/attribute/form/hooks/useMainIdentifierCode');
 
 declare global {
   namespace NodeJS {
@@ -45,6 +46,18 @@ test('it renders a confirm modal delete', async () => {
     5
   );
   expect(screen.getByText('pim_common.confirm_deletion')).toBeInTheDocument();
+});
+
+test('it displays helper when attribute is the main identifier', async () => {
+  await act(async () => {
+    renderWithProviders(<DeleteModal onCancel={jest.fn()} onSuccess={jest.fn()} attributeCode="sku" />);
+  });
+
+  expect(dependencies.translate).toHaveBeenCalledWith(
+    'pim_enrich.entity.attribute.module.delete.cannot_delete',
+    undefined,
+    undefined
+  );
 });
 
 test('it does not allow confirmation until the attributeCodeConfirm field is valid', async () => {
